Docstrings

SwapSort._filename — Method

_filename(N,L,D)

The file that contains the infomation of constructing swapsortN_L_D.

SwapSort.compswap — Method

compswap(i,j)

The expression ai, aj = min_max(ai, aj; kw...)

SwapSort.keywords — Method

keywords()

The expression kw... in a function definition.

SwapSort.localval — Method

localval(i)

The i-th variable name ai.

SwapSort.min_max — Method

min_max(x, y; lt=isless, by=identity) = lt(by(x), by(y)) ? (x,y) : (y,x)

The same as Base.minmax, but supports lt and by keywords which works as those in Base.sort.

SwapSort.returnvals — Method

returnvals(N)

The expression return a1, a2, ..., a(N-1).

SwapSort.swapsort — Method

swapsort(a...; lt=isless, by=identity)

Sort the arguments, returning a sorted tuple. lt and by keywords work the same as Base.sort. swapsort supports up to 64 arguments. See also tuplesort.

SwapSort.tuplesort — Method

tuplesort(t::Tuple; lt=isless, by=identity)

Return a sorted tuple. lt and by keywords work the same as Base.sort. tuplesort calls swapsort and supports at most 64 elements.

SwapSort.varstring — Method

varstring(N)

Returns the string of argument list "a0,a1,...,a(N-1)".